SourceFrameStream: protect against memory loss in WebGL
WebGL is apparently not synced with our calls and renders with an invalid buffer as soon as we let go of the shared memory. And even with this protection, the frontend crashes as soon as it hits listenerFlip. Therefore, I currently copy it into a temporary buffer until this bug can be fixed.