Mixer (#181)
Mixer
TypeScript support for services
Services can emit and listen events
Inject decorator for services
Slider component (extracted from SliderInput)
some more little improvements
About events - I've researched 3 ways: Vue event-bus, EventEmitter, RxJS streams

Vue event-bus works but forced me to write a wrapper that allows to write type-safe code with TS. As it is not a good idea to invent something when we can use something already working and tested I decided to look at other libraries.

EventEmmiter already has some type definitions for TS but it requires to write a lot of boilerplate code to keep type-safe, because events are still strings.

As it turned out the Subject from RxJS was the simplest and shortest way to implement type-safe-events. Fortunately we don't have to import all RxJS library - this library allows to import single classes and functions, and we can use only 2 classes from it - Subject and Subscription. So we don't have big difference with EventEmmiter approach. This library was written by Microsoft on TypeScript and angular2 also use it for events, so we can be sure that we won't have problems with it.

The main difference between Subject from RxJS and EventEmitter - EventEmmiter is an event-bus meanwhile Subject is a typed stream. In contrast to event-bus, stream has finish event that can simplify the code, especially with our node-obs callbacks.

Vue also has an official plugin for RxJS https://github.com/vuejs/vue-rx but for current tasks it is not necessary to use it.