Move scenes to typescript (#229)
It was forced to refactor ScenesService, SourcesService and AudioService to move sources code to TypeScript in the right way that we told before. What do we have now:

Interfaces ISource, ISceneSource, IAudioSource - describes the data-models how it is in stores. It like normalized records in database table.

Classes Source, SceneSource, AudioSource - allow to access the denormalized data in stores via proxifying properties, and allows to declare methods and computed props that will allow us to split the large services code.

SourcesService now doesn't have ScenesService dependency.

This is not the last refactoring changes, I'm gonna try to reduce ScenesService code in the next pull request.