Move rest of services to the new architecture (#481)
Now the all new service will use our new architecture, that means that the all code of services now is executed only in the main window. I checked our services and realized that after implementing promises support and hotkeys refactoring there are only 2 services that unable to work with new way. I've put it to the blacklist.