Rid of blackList in service-manager (#503)
The all code in services and obs-api calls now are executing in the main renderer.
Now WindowService is WindowsService and manages both child and main window. Also the all methods like showSourceProperties now is responsibility of services.
I did some research and found the issue that disallowed us to import WindowService in some services. It was weird circular dependency. Now it should be ok. Also did some cleanup code in this PR.