Optimize properties objects (#49)
Properties objects can take some time to create. As a result, we should
cache them when we can. In addition, I've removed `properties` as a
getter property and changed it to a function to prevent its value from
being enumerated on accident. This should fix several issues including
hitching and freezing.