Modify types and use object spread for more type safety
Seems that the use of `Object.assign` discards type information and
as a result, when we assign to the different InputValue interfaced
objects, it will succeed even if we're not properly assigning according
to the type requirements.

In addition, these changes changed the types to require certain fields
that should always be present.