Update to node-obs integrated with obs-studio-node (#110)
* Update source to match obs-studio-node staging

* Update obs-studio-node to current staging

OSN now exposes node-obs functionality, removing the need for node-obs to be compiled separately. All files are properly shipped by OSN and will be updated with it. This should simplify logic and further advancements in technology concerning obs-studio-node in addition to massively speeding up work on OSN.

* Fix up package.json for node-obs removal

- Removes commands that do this for or with node-obs.
- Updated ignored directories.
- Removed node-obs from extraFiles.

* Use obs-studio-node instead of node-obs

OSN exposes node-obs functionality, so this differentiation is no longer needed.

* vscode: Allow starting with paths containing space

* Update OSN and code to use new node-obs functionality

* Update obs-studio-node

- Displays now load properly.

* ci: AppVeyor should no longer install node-obs

* ci: AppVeyor should automatically install external plugins