Combine ISliderInputForm with INumberInputForm (#160)
When IntInput components were being used, minVal and maxVal
were used while undefined. To fix this as simply as possible,
I combined the ISliderInputForm and INumberInputForm (since
they actually should be the same to begin with) and make sure
any number-based property has a stepVal, minVal, and maxVal.

This fixes any non-slider number property in that it allowed
infinite values. Some plugins didn't like this, such as the
browser source which would crash if a width or height set
a value that was beyond its limits.