modules: Update stream-labs/node-libuiohook (#218)
* modules: Update stream-labs/node-libuiohook

* Switch to "passthrough" hotkey model

This type of key checking is easier to maintain, better supported on other platforms and was already partially in use by the previous code. In addition to that it's also completely free of any delayed errors that may happen, since all the code that does the checking is based on checking manually instead of registering something with the kernel.

Additionally I have removed libuiohook. It has not been used since the implementation of the RegisterHotkey based one and will not be used in the future. This code can practically be merged into obs-studio-node as a utility. If we do merge it with obs-studio-node, we can match OBS's behavior 1:1.

* modules: Update libuiohook with recent fixes

* Fixed the issue where registering hotkeys would immediately trigger them.
* Fixed the issue that caused two hotkeys (Ctrl-A, A) to be triggered at the same time if Ctrl-A is pressed.

Co-authored-by: Eddy Gharbi <eddy.gharbi@streamlabs.com>