Source property optimization (#552)
* Allow passing properties in defaults setup func

Properties is an expensive creation, not made very obvious by the fact
that it's used as a property rather than a function that would probably
be named "createProperties". This fixes an additional call to properties
creation each time a source is made.

* Don't call update unless required

Buttons don't require a source to be updated. Updating a source will
sometimes cause severe context changes and they're only meant to be told
that they were clicked. The button property can do any contextual
changes it wants but the UI shouldn't. This sees if there were any
changes to the property forms that would require updating and chooses
not to update if none were found.

In addition, an optimization was made where if you don't press a button
it won't make an unecessary copy of the source properties.