Cleanup window popup for all windows
Windows are actually slow in Electron. Like... really frickin' slow.
As a result, it's imperative that we make sure we're done rendering
before we actually show the window else the user will be presented with
all sorts of garbage.

For the main window, turns out we don't seem to wait for the
`ready-to-show` event in addition to not setting a background color.
So this should get rid of the white screen on load and if we still
somehow take to long to render in time for the `ready-to-show` event,
we should get a nice solid background that isn't a plain white but more
in line with our application's colors.

DO NOT MERGE RIGHT NOW. THIS IS EXPERIMENTAL.