Obs renderer (#906)
* project: Directly access OBS in Renderer processes

As we now have our own IPC framework, we no longer have to rely on electron's IPC framework to do our work for us. We can directly call things from any process, as long as that process is actually connected to the IPC server.

For OBS IPC, this means a lot of data can be directly accessed without blocking both the renderer and main process, thus freeing up resources for more responsiveness and faster rendering. As a side effect, we might see lower CPU usage too.

* app: Ensure environment variable is valid

* obs-api: Fix formatting

* main: Fix working directory and path.join

* services/app: Fix shutdown crash and ensure env variable is valid