Video and Image File Compression
Ran a series of lossless compression algorithms to make the videos and images smaller. Also remuxed .mp4 files in H-264 in VLC. Total reduction of ~11,580,000 bytes.

Logs available upon request... doing this for a bunch of projects, and I'd have to grep the images for this work in particular. Example here: https://pastebin.com/MTtcVck2