Video and Image File Compression - attempt #2
Ran a series of lossless compression algorithms to make the videos and images smaller. Also remuxed .mp4 files in H-264 in VLC. Total reduction of ~11,580,000 bytes.

Logs of tasks performed available upon request... doing this for a bunch of projects, and I'd have to grep the images for this work in particular. Example here: https://pastebin.com/MTtcVck2

Repull of this: https://github.com/stream-labs/streamlabs-obs/pull/1041