feat(platform): create scene transitions from Platform apps
Add a new module to Platforms API, Scene Transitions, responsible for
providing an API for platform developers to create scene transitions
programmatically.

Currently, only stinger transitions are supported, though the facade
involved in the translation between a developer-friendly API and what
OBS expects it's pretty generic.  The video files used in stinger
transitions are expected to be provided as part of the app (i.e a
relative path).

There's an option for app developers to lock specific scene transitions,
preventing the user from editing them. Changes to the Manage Scene
Transitions dialog were implemented accordingly.

This `createTransition` method takes developer-friendlier list of
params, simplifying the use of this API. This is very generic, in order
to support more transitions in the future, and yet the details are
encapsulated from the developer. This commit also retrieves the file
from the application path, treating it as an asset as opposed to
previous implementation which relied on hosted URLs.