OBS as source of truth for hotkeys (#1054)
* refactor: OBS hotkeys

* support display capture hotkeys.
* fix issues with duplicate scenes.
* support sending keyup events to OBS backend.
* log error when scene collection fails to load.

Now we can process `SHOW_SCENE_ITEM` and `HIDE_SCENE_ITEM` without
getting duplicate scenes in `SceneSelector`, coming from the service.
An uniqueness check was added to the getter for scenes that checks
based on scene ID.

If we fail to load an scene collection, the behavior appears to be
trying to restore from backup. This causes the loading of application
state to be invoked a second time, with unexpected side-effects like
duplicate scenes. This commit addresses the visibility part by
displaying an error message for developers as opposed to swallowing the
error entirely. Prior to this, we addressed the second part of one of
those visible effects by adding uniqueness checks in the scenes API.

* Fortifies checking for a source's type in case it cannot be found in
the service (it has happened).
* Fix action names for `TOGGLE_MUTE` and `TOGGLE_UNMUTE`.
* Track added hotkeys on a set for the time being, ensuring we don't
add duplicates coming from OBS that we're already handling.
* Add explicit check for whether the returned item belongs to a scene
hotkey.

Please notice that currently identifying which kind of hotkey we got
from the backend is factorial as we don't seem to have a better way to
get this info from the backend. The number of hotkeys affected by this
is minimal (the main reason we're sticking to the former loops for now)
but could still be significant. I have not, however, profiled this.

* chore: add comment about dup items

* refactor: rename indexer to actionName