feat(troubleshooter): add control to start/stop streaming
In order to be able to change the settings inline while you're streaming we need to provide the
ability to end the stream from the window itself, otherwise the controls are useless unless the user
is accessing the notification at a later time when they're offline.

This commit also adds emphasis to the message that the stream needs to be stop to access these
controls.