fix(windows): creating window bigger than current display res, style fixes
Override window size to avoid creating a window bigger than the current display.
We do not exercise this behavior on CI as it runs at 1024x768 and breaks tests that aren't easy to
workaround.

Fixed widths seem to have no effect on high resolutions, but in low resolutions it makes the
"Selected Properties" section of the Widgets Editor to have invisible or out of place content,
including the scrollbar, color picker, and probably others.

I've added media queries to fix the particular case of low resolution, but we should consider
removing the fixed widths.