feat(twitch): tags support (#1042)
* day and night mode theme support.
* editing tags while the stream is live.
* fetch active tags for a given stream.
* update Twitch endpoints.
* update tags code to new API.
* ability to retrieve tags for a particular stream.
* remove need to query streams to update tags.
* `Helix` API requires `Bearer` as auth token prefix.
* retrieve all tags after Twitch pagination fix.
* twitch tags control and handling
* initial support for tags and streams APIs.
* provide `afterGoLive` hook, details below.

Provide a hook that could be invoked by platform services, after finishing
going live. We're notified of live status through signal handling, we chose to
invoke such hook directly if exists, as opposed to letting the client
subscribe, since there are many layers of indirection.

`startStreaming` and `toggleStreaming` now accept an optional context, which is
passed to the aforementioned `afterGoLive` hook. This is necessary if the
caller wants to use some data that was passed prior to the stream starting.

refactor: better handling by using stream info

refactor: prefix types