fix: scene transitions API after deletion
After deleting a scene transition through the UI, if the developer creates a transition pointing to
the same file, the transition would fail to run as we failed to update the path to the file on disk,
since we didn't need to download the file again (as it was on our asset link/checksum) we submitted
the URL as-is to OBS which is not what we want. We update the URL to the file on disk if that were
to be the case.

Reproduction case given:

1. Create a transition via API.
2. Set as default.
3. Change scenes (it'd work)
4. Delete transition through SLOBS UI.
5. Create the same transition through API.
6. Set as default
7. Switch scenes (would no longer work).