fix(inputs): exception when opening streamlabel source
Opening a streamlabel source and selecting "Use Google Font" triggers an exception as this
components uses a `loading` property that appears to conflict with our new inputs / form handling,
as a result, Vue thinks is the parent prop we're mutating and besides a warning the general results
could be unreliable. Fix just renames prop to `isLoading`.