fix(inputs): exception when opening streamlabel source (#1413)
Opening a streamlabel source and selecting "Use Google Font" triggers an exception as this
component uses a `loading` property that appears to conflict with our new inputs/form handling, as a result, Vue thinks is the parent prop we're mutating and the general handling of that could be unreliable. Fix just renames prop to `isLoading`.