ci: remove upsource triggers
As Upsource is not being used at the moment, remove the steps necessary to make the CI integration
work, namely a custom step that established the commit ID (which is now only adding noice to PR
comments), and the webhook (which is definitely failing, and consuming a few seconds).

We can look back at this commit to re-enable, if we ever choose so.