Overlay v1 (#1744)
* feat: game overlay (WIP)

* feat(overlay): see details

* fix recent events page wouldn't load
* hotkey bindings, offscreen rendering wip
* scene switcher, streaming button, refactor, etc
* refactor imports, extra els, etc
* overlay settings
* fix spaces are no longer required for thread status
* refactor subscription order, display method
* draggable windows, refactors, details follow

* Draggable windows: without titlebar windows are not draggable, this commit
injects CSS inparticular areas to restore draggable behavior. On Recent Events
blank space around the navigationbar, for Chat the left-side section next to
the room selector, for overlay controls the entirewindow except for clickable
elements. Draggable areas are generally more lenient, the previous is
an approximation. To test this before get input interception we need to use
`show` as opposed to`showInactive` and reset `OFFSCREEN_OFFSET` to 0.
* `onWindowsReady` is now dynamic and now listens for the amount of windows in
`this.windows` asopposed to fixed number.
* Enable `contextIsolation` for views.
* Remove BrowserView and `loadURL` directly, this was required for JavaScript
injection (didn't seem to work  with BrowserView's) and it doesn't seem to have
any issues. Also adjusts window bounds accordingly (they're still positioned to
look like a single window unless you start dragging them around).
* More general refactors.

* login lifecycle for actions

Abstract a common pattern of performing an action if logged in, subscribing to user login events to
perform that action if the user wasn't logged in, and cleanup on logout.

* dragging recent event sections, refactor
* optimize imports
* only display toggle hotkey if enabled
* add game overlay translations
* optimize imports on overlay controls win
* remove common browser view options
* re-enable `contextIsolation` for views

As we moved away from `BrowserView` we removed their preferences and `contextIsolation` was not
added to the respective `BrowserWindow`'s. While we control one of sources (streamlabs.com) and the
other one is Twitch Chat, there shouldn't be any issues, but this protects us regardless.

* remove reload workaround for overlay controls

Now that windows are initially rendered offscreen it doesn't matter whether we get an initial white
flash on the controls window, by the time the overlay is shown it should've correctly captured the
windows' content.

TODO: ensure we don't have race conditions.

* make all views draggable at once, dynamic [ci skip]
* refactor app loading lifecycle and destroy [ci skip]

* Fix merge regressions

* Update game-overlay

* Fix merge regression

* Update overlay

* Fix merge conflicts

* Switch to paintOverlay api

* show overlays

* Remove overlayControls until interactivity is ready

* Add move-able windows

* Add in full settings

* Refactor overlay service

* Fix tests

* Fix opacity not initializing improperly

* Remove hotkeys from overlay options window

* Improve preview mode ux

* Add styles to placeholder

* Fix spawning too many preview windows

* Fix overlayId conditional

* Minor refactor

* Address CC

* Add position saving and reset button

* Save previewWindow position properly

* Update game-overlay

* Restrict resizing

* Remove deprecated file from gitignore

* Address CR

* Remove parent window