Improve JSON-RPC implementation (#3570)
* Improve type validation of JSON-RPC ids

According to JSON-RPC 2.0 the ids of requests and
responses can be strings, numbers or `null`. This commit
updates the types of ids and improves the response,
event and error creation functions.

Additionally, this commit removes the
`eslint-disable`-rules in JsonRpcService.

For less validation and errors it unifies the arguments
passed to the functions as request objects instead of
request objects or request ids. Since the request is
sometimes optional, an argument reordering is necessary
causing further changes in other functions.

* Fix wrong function call in createEvent

Fix wrong function call in `createEvent`.
Function `createEvent` is not used, therefore
the bug does not occur somewhere yet.