Resolve spin wheel button not showing up on mini feed when widget is added to scene (#4419)
* Update call to a reactive get. Update other instances calling function

* Update to WidgetsService/WidgetsServiceViews based upon request

Brought up the widgetSources and widgetSource up to WidgetsServiceViews. There is still a bit of cleanup that will need to happen as we further deprecate.

Tested widgets and chat based upon the changes.

* "Fix" for obs-importer test.

This won't be its final form. Something on the WidgetServiceViews isn't being initiated correctly causing the failure.

Co-authored-by: Thom Swann <tswann@logitech.com>