fix(facebook): event doesn't appear on list after created until reload
Newly-created `LiveVideo`'s do not seem to have `broadcast_start_time`,
which causes one of the many conversion logics around them to fail and
result in an invalid date, preventing the newly-created event from
showing on the list until the view is reloaded and data is fetched and
converted again.

NOTE: there's still a bug on the scheduler, clicking any event does not
show the Edit modal, it fails to fetch the proper video/event
information (they're a mix at this point), revolving around
`fetchVideo` and `fetchStartStreamOptionsForVideo`.