fix(hotkeys): incorrect length calc caused missing source hotkeys
`sourceHotkeys` is an object where I was using the wrong length
calculation method on, as a result I believe it was always set
to true and defaulted to a single render with the original OBS
sources (Desktop Audio, etc) but not the sources added to scenes.

Also hides the Markers title if there aren't any markers.