fix(dual-output): missing Filters, Rename, Properties from ctx menu
The "Filters", "Rename" and "Properties" options on the context menu
when right-clicking on a source on the Editor is missing. This is due
to selection having multiple elements and thus triggering the
multiselection condition which hides these items. What really happens is
that the selection is comprised of two nodes (one for each display), and
while this behavior only seems to happen when selecting a source from the
source pane, it has been reported to happen throughout the editor (I
can't verify).

Refactoring the Selection service is a more convuluted task so we've
followed a similar approach to what the Edit Transform submenu item does
but just to verify its condition (2 nodes but really 1 node across two
displays), the 3 commands in question seem to grab a single selection
item which should be one on the last display selected (gathered from the event)
so while we need to test more we noticed these commands are working
correctly (on said display).

ref: https://app.asana.com/0/734380881425048/1205635700788639/f