fix(streaming): platform disabled switch, details follow
Setting the `Switch` itself to include the check for a platform being
disabled makes it so that you can't enable an initially disabled
platform through it. With the context that I have, which is minimal, any
checks that require the old functionality should probably be done before
this gets rendered.