fix(layout): attempt to fix layout editor issue, details follow
We seem to be navigating before setting the LayoutEditor's current tab,
which would probably mean the order is wrong or something in Editor at
the very root is not reactive enough on `LayoutService.currentTab` to
cause a re-render. As a result, clicking on a new layout editor tab
from the sidebar shows the previous current tab before the call to
`setCurrentTab` seems to take any effect.