fix(multistreaming): filter out platforms that don't have chat
With Tiktok being part of the active platforms seemingly at all times,
will make `PrimaryChatSwitcher` blow up since it doesn't technically
support chat. We have refactored to make this explicit in code, by
not only making sure that we only pass chat-enabled platforms to this
component but also that we hide it if there's no other platform aside
from the primary that could be set for primary chat. That is such the
case for example when using Twitch as primary and also streaming to
Tiktok.