fix(chat): TikTok's fake chat refactor
TikTok's fake chat component makes other platform chats unable to load
after switching to Tiktok's chat while live. This simplifies Tiktok's
handling by moving logic to the parent component and not rendering a
`Chat` component at all since we don't need to load an URL or setup
resize event listeners, or any of the other functionality we're aware
of. While the code might not be pretty on the caller, getting rid of
render side effects and unrelated code in the `Chat` component that's
already been found to be buggy should be an improvement overall.