fix: invalid recording path error swallow (#5036)
If we don't get JSON from OBS error we swallow some errors like an
invalid recording path and show a generic "Unknown Error" message.

OBS still treats this as a `-4` error code, which makes it take this
branch. Try to improve this by attempting to parse the JSON otherwise
return a JSON object resembling what the original code expects.