fix(collectibles): ensure media dir exists, details follow
On a fresh cache, where the user has no files that have triggered media
backup before, this directory does not exist, so attempting to download
a collectible to that directory fails. This ensures the directory exists
prior to downloading. Unfortunately we had to make that method public
since we're not using `MediaBackupService.downloadFile` here.