fix(collectibles): not found scenes did not throw
Somewhere we regressed and not found scenes were no longer throwing an
error but returning undefined silently due to the `?` operator.

Refactors creation method out of the view model, while something else
was causing issues earlier, it no longer seems to be the case and we can
correctly access the scene returned from `this.views.getScene()` on the
service, what remained was a serialization error, which we removed by
returning `sceneItemId` from it, as our API defines.