feat: add collectibles API for adding source inside embed (#5135)
* feat: add collectibles API for adding source inside embed

Adds guest API methods for listing scenes and adding a collectible
to a specific scene. This is convenience for create source depending
on its type.

* feat(collectibles): add boolean for active scene

* fix(collectibles): add staging URL to validation

* fix(collectibles): invert boolean logic

* feat(collectibles): require passing source name

* feat(collectibles): start with file download

* refactor(collectibles): very minor cleanup

* fix(collectibles): incomplete URL sanitization in check

Even though we control domain and caller, I'd rather do it properly.
It seems like the overlay host validation uses the same method.

* fix(collectibles): ensure media dir exists, details follow

On a fresh cache, where the user has no files that have triggered media
backup before, this directory does not exist, so attempting to download
a collectible to that directory fails. This ensures the directory exists
prior to downloading. Unfortunately we had to make that method public
since we're not using `MediaBackupService.downloadFile` here.

* fix(collectibles): not found scenes did not throw

Somewhere we regressed and not found scenes were no longer throwing an
error but returning undefined silently due to the `?` operator.

Refactors creation method out of the view model, while something else
was causing issues earlier, it no longer seems to be the case and we can
correctly access the scene returned from `this.views.getScene()` on the
service, what remained was a serialization error, which we removed by
returning `sceneItemId` from it, as our API defines.

* feat(collectibles): add vertical nodes