fix(volmeters): missing bars on app load or device switch
When Desktop Audio (and other sources presumably) haven't played audio
we don't get bars in the mixer until some app starts playing audio to
it, as a result on many scenarios on app load or source device switch
we'd get missing volmeter bars with just the source name which looks
janky. I'd like to revisit the rendering of volmeters at some point
entirely, since first impressions seems like it's drawing several times
a second regardless of whether events are received or not and, in this
case, not even rendering at all if none are received, but for now this
is a hack that sets a valid peak on every new subscription to ensure we
draw at least once.