fix(collectibles): only create partner nodes if dual output enabled
Unconditionally creating partner nodes when dual output is disables
makes the source appear twice in the editor, due to that vertical node
being created. Side effect of this is that if the user adds a
collectible and only then enables dual output, they'll probably have to
wait till a full app launch for it to show on the vertical display.