fix(widgets): custom code in widgets have no defaults so they break (#5202)
Widgets can get into an invalid state when enabling custom code, no defaults were consistently provided by backend and, as a result, enabling the Custom Code option could result in empty widget HTML, CSS and JS, thus breaking the widget.
In some cases, users could recover by going through the dashboard and choosing to "Reset Defaults", but this change makes that invalid state impossible.
We fetch custom code defaults through the new static config API at `/api/v5/widgets/static/config/{widgetType}` and populate those fields if they're empty on widget settings load. 
We needed to add a new `humanType` field which is the widget type in string representation as opposed to the number enum so we can pass it to that endpoint.
Also updates the `custom_defaults` property that is supposed to be used to provide that reset functionality on Desktop, though is inconsistent where we've seen this button be available.
Backend is working towards a complete fix which will return the default code in the standard settings endpoints if they're empty to make sure it never breaks again, and also that users can't end up in invalid state by resetting them through the web UI.